Allow for needing release asset as download package
Gutenberg and other plugins may require a build process. This means a different download link will be need which doesn't nest the repo.

The normal download link is `https://api.gitbhub.com/repos/<owner>/<repo>/zipball/<tag>`

There is a new config variable `release_asset` that is a boolean and needs to be set accordingly.